factor out code publishing logic
