+++ api/src/tests/view_function.rs
@@ -4,8 +4,6 @@
 use super::new_test_context;
 use aptos_api_test_context::current_function_name;
 use aptos_cached_packages::aptos_stdlib;
-use aptos_framework::{BuildOptions, BuiltPackage};
-use aptos_package_builder::PackageBuilder;
 use serde_json::json;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
@@ -85,31 +83,16 @@ async fn test_versioned_simple_view() {
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn test_view_tuple() {
     let mut context = new_test_context(current_function_name!());
-
-    let mut builder = PackageBuilder::new("test_package");
-    builder.add_source(
-        "test_module.move",
+    let payload = aptos_stdlib::publish_module_source(
+        "test_module",
         r#"
         module 0xa550c18::test_module {
-
             #[view]
             public fun return_tuple(): (u64, u64) {
                 (1, 2)
             }
         }
         "#,
-    );
-    let path = builder.write_to_temp().unwrap();
-
-    let package = BuiltPackage::build(path.path().to_path_buf(), BuildOptions::default())
-        .expect("Should be able to build a package");
-    let code = package.extract_code();
-    let metadata = package
-        .extract_metadata()
-        .expect("Should be able to extract metadata");
-    let payload = aptos_stdlib::code_publish_package_txn(
-        bcs::to_bytes(&metadata).expect("Should be able to serialize metadata"),
-        code,
     );
 
     let root_account = context.root_account().await;
