+++ aptos-move/framework/cached-packages/Cargo.toml
@@ -14,6 +14,7 @@ rust-version = { workspace = true }
 
 [dependencies]
 aptos-framework = { workspace = true }
+aptos-package-builder = { workspace = true }
 aptos-types = { workspace = true }
 bcs = { workspace = true }
 include_dir = { workspace = true }
