+++ aptos-move/framework/cached-packages/src/aptos_stdlib.rs
@@ -7,6 +7,8 @@ pub use crate::{
     aptos_framework_sdk_builder::*, aptos_token_objects_sdk_builder as aptos_token_objects_stdlib,
     aptos_token_sdk_builder as aptos_token_stdlib,
 };
+use aptos_framework::{BuildOptions, BuiltPackage};
+use aptos_package_builder::PackageBuilder;
 use aptos_types::{account_address::AccountAddress, transaction::TransactionPayload};
 
 pub fn aptos_coin_transfer(to: AccountAddress, amount: u64) -> TransactionPayload {
@@ -15,4 +17,20 @@ pub fn aptos_coin_transfer(to: AccountAddress, amount: u64) -> TransactionPayloa
         to,
         amount,
     )
+}
+
+pub fn publish_module_source(module_name: &str, module_src: &str) -> TransactionPayload {
+    let mut builder = PackageBuilder::new("tmp");
+    builder.add_source(module_name, module_src);
+
+    let tmp_dir = builder.write_to_temp().unwrap();
+    let package = BuiltPackage::build(tmp_dir.path().to_path_buf(), BuildOptions::default())
+        .expect("Should be able to build a package");
+    let code = package.extract_code();
+    let metadata = package
+        .extract_metadata()
+        .expect("Should be able to extract metadata");
+    let metadata_serialized =
+        bcs::to_bytes(&metadata).expect("Should be able to serialize metadata");
+    code_publish_package_txn(metadata_serialized, code)
 }
